feat(Migration V2 V3): move sample into its own file
